+++ src/pytest_djangoapp/fixtures/utils.py
@@ -1,7 +1,8 @@
 from time import time
 
 import pytest
-from django.utils import translation
+from django.urls import get_resolver
+from django.urls.resolvers import URLResolver
 
 from ..configuration import Configuration
 
@@ -11,7 +12,45 @@ def get_stamp() -> str:
     return f'{time()}'
 
 
-@pytest.fixture(scope='session')
+@pytest.fixture()
+def registered_urls() -> dict:
+    """Returns mapping from registered urlpatterns to names and args.
+
+    ```py
+    def test_me(registered_urls):
+        assert 'admin:testapp_article_change object_id' in f'{registered_urls}'
+    ```
+
+    """
+    patterns = {}
+
+    def clean_url(prefix, postfix) -> str:
+        prefix = f'{prefix}'.strip(' ^$')
+        postfix = f'{postfix}'.strip(' ^$')
+        return f'{prefix}{postfix}'.replace('\\\\', "\\")
+
+    def gather(resolver, *, url_prefix: str, parent_namespace: str, container: dict):
+        namespace = resolver.namespace or parent_namespace
+        url_prefix = clean_url(url_prefix, resolver.pattern)
+
+        for item in resolver.url_patterns:
+            subcontainer = container.setdefault(namespace, {}) if namespace != parent_namespace else container
+
+            if isinstance(item, URLResolver):
+                gather(item, url_prefix=url_prefix, parent_namespace=namespace, container=subcontainer)
+            else:
+                pattern = item.pattern
+                args = ' '.join(pattern.converters)
+                subcontainer[clean_url(url_prefix, pattern)] = (
+                    f'{namespace}:{pattern.name or "???"} {args}'.strip(': ')
+                )
+
+    gather(get_resolver(), url_prefix='', parent_namespace='', container=patterns)
+
+    return patterns
+
+
+@pytest.fixture(scope='session')#
 def conf_app_name() -> str:
     """Returns your application name (deduced or provided)."""
     return Configuration.get_dict()[Configuration.KEY_APP]