+++ tests/test_fixtures_utils.py
@@ -6,6 +6,10 @@ def test_conf_app_name(conf_app_name):
     assert conf_app_name == 'pytest_djangoapp'
 
 
+def test_registered_urls(registered_urls):
+    assert 'admin:testapp_article_change object_id' in f'{registered_urls}'
+
+
 def test_run_app(run_app, user_model):
 
     with patch('django.core.management.commands.runserver.Command.run') as patched: